refactor hero & random hero
