+++ src/commands/mod.rs
@@ -1,4 +1,3 @@
 pub mod hero;
-pub mod randomhero;
 pub mod team;
 pub mod teams;
