+++ src/events.rs
@@ -2,7 +2,7 @@ use std::env;
 
 use serenity::{
     async_trait,
-    model::prelude::{command::CommandOptionType, interaction::Interaction, GuildId, Ready},
+    model::prelude::{interaction::Interaction, GuildId, Ready},
     prelude::{Context, EventHandler},
 };
 
@@ -17,8 +17,8 @@ impl EventHandler for Handler {
             Interaction::ApplicationCommand(ref command) => match command.data.name.as_str() {
                 "team" => commands::team::run(ctx, command.to_owned()).await,
                 "teams" => commands::teams::run(ctx, command.to_owned()).await,
-                "hero" => commands::hero::run(ctx, command.to_owned()).await,
-                "randomhero" => commands::randomhero::run(ctx, command.to_owned()).await,
+                "hero" => commands::hero::hero(ctx, command.to_owned()).await,
+                "randomhero" => commands::hero::random_hero(ctx, command.to_owned()).await,
                 other_commands => println!("Unknown command {}", other_commands),
             },
             Interaction::MessageComponent(ref component) => match component.message.interaction {
@@ -59,22 +59,9 @@ impl EventHandler for Handler {
             commands.create_application_command(|command| commands::team::register(command));
             commands.create_application_command(|command| commands::teams::register(command));
             commands.create_application_command(|command| {
-                command
-                    .name("randomhero")
-                    .description("Get a random contributor to novu's repository")
+                commands::hero::register_random_hero(command)
             });
-            commands.create_application_command(|command| {
-                command
-                    .name("hero")
-                    .description("Get info on a novu community hero")
-                    .create_option(|option| {
-                        option
-                            .name("github_username")
-                            .description("The github username of the contributor to look for")
-                            .kind(CommandOptionType::String)
-                            .required(true)
-                    })
-            })
+            commands.create_application_command(|command| commands::hero::register_hero(command))
         })
         .await;
